+++ hw-ledger/types.ts
@@ -1,3 +1,5 @@
+import { Buffer } from 'node:buffer';
+
 
 import type { HexString } from 'https://deno.land/x/polkadot/util/types.ts';