+++ util/u8a/toHexBuffer.ts
@@ -1,3 +1,5 @@
+import { Buffer } from 'node:buffer';
+
 
 import type { HexString } from '../types.ts';