+++ x-fetch/packageInfo.ts
@@ -1,3 +1,3 @@
 
 
-export const packageInfo = { name: '@polkadot/x-fetch', path: new URL(import.meta.url).pathname, type: 'deno', version: '13.4.2' };
+export const packageInfo = { name: '@polkadot/x-fetch', path: new URL(import.meta.url).pathname, type: 'deno', version: '13.4.3' };